requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

# Infrastructure
oslo.utils == 2.2.0
oslo.messaging == 2.2.0
oslo.config == 2.2.0
oslo.i18n == 2.3.0
oslo.log == 1.8.0
oslo.service == 0.6.0
oslo.concurrency == 2.4.0
oslo.middleware == 2.5.0

pbr == 1.4.0
Babel == 2.0
eventlet == 0.17.4
Paste == 2.0.2
PasteDeploy == 1.5.2
paramiko == 1.15.2
webob == 1.4.1
Routes == 2.2

# 1st step Authentication
keystonemiddleware == 2.1.0

# Openstack Client APIs
python-novaclient == 2.26.0

# Image Generation
docker-py == 1.3.1
